Create starcluster launch script and incorporation into awsls page

+++ kill_instance.py
@@ -4,8 +4,8 @@ import subprocess
 import os
 import sys 
 if len(sys.argv) ==1:
-	print '\nUsage: awskill [instance ID]\n'
-	print '\nSpecify instance ID that will be terminated, which can be found using "awsls"\n'
+	print '\nUsage: awskill [instance ID or cluster name]\n'
+	print '\nSpecify instance ID or cluster name that will be terminated, which can be found using "awsls" or "awsls -c"\n'
 	sys.exit()
 
 instanceID=sys.argv[1]
@@ -46,17 +46,60 @@ if keyPath.split('/')[-1].split('.')[-1] != 'pem':
 
 tag=keyPath.split('/')[-1].split('.')[0]
 
-answer=query_yes_no("\nTerminate instance %s?" %(instanceID))
+starcluster=subprocess.Popen('which starcluster',shell=True, stdout=subprocess.PIPE).stdout.read().strip()
+if len(starcluster) == 0:
+	clusterflag=0
+if len(starcluster) > 0: 
+	clusterflag=1
+if not os.path.exists('%s/.starcluster' %(subprocess.Popen('echo $HOME',shell=True, stdout=subprocess.PIPE).stdout.read().strip())):
+        os.makedirs('%s/.starcluster' %(subprocess.Popen('echo $HOME',shell=True, stdout=subprocess.PIPE).stdout.read().strip()))
 
-if answer is True:
+if not os.path.exists('%s/.starcluster/config' %(subprocess.Popen('echo $HOME',shell=True, stdout=subprocess.PIPE).stdout.read().strip())):
+        AWS_ACCESS_KEY_ID=subprocess.Popen('echo $AWS_ACCESS_KEY_ID',shell=True, stdout=subprocess.PIPE).stdout.read().strip()
+        AWS_SECRET_ACCESS_KEY=subprocess.Popen('echo $AWS_SECRET_ACCESS_KEY',shell=True, stdout=subprocess.PIPE).stdout.read().strip()
+        AWS_ACCOUNT_ID=subprocess.Popen('echo $AWS_ACCOUNT_ID',shell=True, stdout=subprocess.PIPE).stdout.read().strip()
+        AWS_DEFAULT_REGION=subprocess.Popen('echo $AWS_DEFAULT_REGION',shell=True, stdout=subprocess.PIPE).stdout.read().strip()
+        cmd='####################################\n'        
+	cmd+='## StarCluster Configuration File ##\n'
+        cmd+='####################################\n'        
+	cmd+='[aws info]\n'
+        cmd+='AWS_USER_ID=%s\n' %(AWS_ACCOUNT_ID)
+        cmd+='AWS_ACCESS_KEY_ID =%s\n' %(AWS_ACCESS_KEY_ID)
+        cmd+='AWS_SECRET_ACCESS_KEY = %s\n' %(AWS_SECRET_ACCESS_KEY)        
+	cmd+='AWS_REGION_NAME = %s\n' %(AWS_DEFAULT_REGION)
+        cmd+='AVAILABILITY_ZONE = %s\n' %(params['zone'])
+        cmd+='AWS_REGION_HOST = ec2.%s.amazonaws.com\n' %(AWS_DEFAULT_REGION)
+        cmd+='[global]\n'
+        cmd+='DEFAULT_TEMPLATE=cluster\n'
+        cmd+='[key %s]\n' %(keyPath.split('/')[-1].split('.')[0])
+        cmd+='KEY_LOCATION=%s\n' %(keyPath)
 
-	print '\nRemoving instance ...\n'
+	o1=open('%s/.starcluster/config' %(subprocess.Popen('echo $HOME',shell=True, stdout=subprocess.PIPE).stdout.read().strip()),'w')
+        o1.write(cmd)
+        o1.close()
 
-	if os.path.exists('tmp4949585940.txt'):
-		os.remove('tmp4949585940.txt')
+if instanceID.split('-')[0] == 'cluster':
 
-	cmd='aws ec2 terminate-instances --instance-ids %s > tmp4949585940.txt' %(instanceID)
+	if clusterflag==0: 
+		print 'Error: Could not find starcluster installed. Exiting.\n'
+		sys.exit()
+
+	cmd='starcluster terminate %s -f'%(instanceID)
 	subprocess.Popen(cmd,shell=True).wait()
 
-	os.remove('tmp4949585940.txt')
+if instanceID.split('-')[0] != 'cluster':
+
+	answer=query_yes_no("\nTerminate instance %s?" %(instanceID))
+
+	if answer is True:
+
+		print '\nRemoving instance ...\n'
+
+		if os.path.exists('tmp4949585940.txt'):
+			os.remove('tmp4949585940.txt')
+
+		cmd='aws ec2 terminate-instances --instance-ids %s > tmp4949585940.txt' %(instanceID)
+		subprocess.Popen(cmd,shell=True).wait()
+
+		os.remove('tmp4949585940.txt')
